requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
attrs==25.4.0
billiard==4.2.3
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
cryptography==46.0.2
defusedxml==0.7.1
Deprecated==1.2.18
dj-database-url==3.0.1
dj-rest-auth==5.0.1
Django==6.0
django-anymail==13.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg==3.2.10
psycopg-binary==3.2.10
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pyngrok==7.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
python3-openid==3.2.0
PyYAML==6.0.3
redis==7.0.1
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
resend==2.17.0
rpds-py==0.29.0
sendgrid==6.12.5
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
svix==1.77.0
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20251008
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
whitenoise==6.11.0
wrapt==1.17.3
